updated functionality of progress counter.
