+++ mosaicizers/ProgressCounter.py
@@ -1,0 +1,9 @@
+class ProgressCounter:
+    def __init__(self, total):
+        self.total = total
+        self.counter = 0
+
+    def update(self):
+        self.counter += 1
+        print("Progress: {:04.1f}%".format(100 * self.counter / self.total),
+              flush=True, end='\r')
